Fix FASTA parameter not being set properly
Parameters referenced in workflows are **fixed** at the time of
import, and can only be overridden with CLI. This works with 99% of
the parameters, *except* iGenomes which are defined before.

This workarounds the issue so that it will work, but more substantial
changes are required to handle this properly. But at least it works.

Fixes #19.