Makre sure PoN building works with data without chrX
This version of ichorCNA is old and still relies on the old scripts
provided by the package. createPanelOfNormals however assumes that
there is always coverage on chromosome X, and if NULL, tries to
match it with reads on chrY (NA) therefore causing a cryptic error.

Patch the script so that it sets coverage to NA for chrX if missing.